Move project target widgets to Views/ and rename Widget → Form
Three things in one commit:

1. `Components::ProjectTargetWidgetBase` moves to
   `Views::Controllers::Projects::TargetWidgetBase`. It's the shared
   Pattern-B parent for the two inline "add target X" forms — used
   only by the two subclasses, both of which now live under
   `Views::Controllers::Projects::*`. Belongs alongside its only
   callers, not in components/.

2. `Components::ProjectTargetNamesWidget` moves to
   `Views::Controllers::Projects::TargetNames::Form`. The
   target_names POST/DELETE actions live in
   `Projects::TargetNamesController`, so the form file now sits
   under the matching `projects/target_names/` view directory.
   Renamed from "Widget" to "Form" — "Widget" carried no semantic
   load; matches the single-form-per-controller `Form` convention
   used elsewhere (e.g. `Views::Controllers::Names::Trackers::Form`).

3. Existing
   `Views::Controllers::Projects::Locations::TargetLocationsWidget`
   already in views/ — symmetrized to
   `Views::Controllers::Projects::TargetLocations::Form` to match
   the new TargetNames placement. The locations widget previously
   lived under `projects/locations/` (its host page), but the
   form's write owner is `Projects::TargetLocationsController` —
   a real, separate CRUD controller for the
   `project_target_locations` join table, distinct from
   `Projects::LocationsController#index` (read-only rollup of
   observation locations). Form file follows the write owner.

The rendered DOM ids (`target_names_widget`, `target_locations_widget`)
are kept via each subclass's `dom_id` override — both are referenced
by Turbo-stream `replace` calls and SCSS selectors in
`_form_elements.scss`.
