+++ app/controllers/projects/target_names_controller.rb
@@ -55,7 +55,7 @@ module Projects
     # Parse name input: supports comma-separated, newline-separated,
     # and pasted checklist format with counts like "Name (3) * +".
     # Submitted under the `project_target_names_add` namespace by
-    # Components::ProjectTargetNamesWidget (Pattern B form).
+    # Views::Controllers::Projects::TargetNames::Form (Pattern B form).
     def parse_names_from_params
       raw = params.dig(:project_target_names_add, :names).to_s
       raw.split(/[,\n]/).filter_map { |entry| lookup_name(entry) }
